requirements.txt
Flask
Flask-SocketIO
openai
Pillow
reportlab
requests
Werkzeug
gunicorn
python-dotenv
python-socketio
python-engineio
better-profanity
opencv-python
numpy
Flask-SQLAlchemy
psycopg2-binary
alembic
Flask-Migrate
Flask-Login
sqlalchemy
gevent
gevent-websocket